requirements.txt
google-generativeai
python-dotenv
appwrite
